% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{gpSIM}
\alias{gpSIM}
\title{Linear Model for Genome-wide Association Studies}
\usage{
gpSIM(nIDs, nSNPs, h2, nQTL, CatEff = TRUE, nCat = 5, CovEff = TRUE)
}
\arguments{
\item{nIDs}{Number of individuals.}

\item{nSNPs}{Number of markers.}

\item{h2}{Heritability.}

\item{nQTL}{Number of QTLs.}

\item{CatEff}{A logical value indicating whether a categorial variable must be used.}

\item{nCat}{Number of distinct categories in CatEff.}

\item{CovEff}{A logical value indicating whether a covariate variable must be used.}
}
\value{
A text file with GWAS statistics. Manhattan plot and QQ-Plot. If an interaction 
        between the SNP and any other fixed effect is in the model, its pvalues and 
        some statistics are also saved in the text file.
}
\description{
The lmGWAS (Linear Model for Genome-wide Association Studies) is developed 
             as a package for GWAS based on a single SNP analysis .
}
